Desktop/7755/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.6.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.14.0
requests==2.31.0
